Automate file import
Combined the code for a single, multi-dive annotation file with the code for a group of dives saved as a batch of .csv files using if/else. If multiple files exist in the annotation folder, it will execute the batch import option, otherwise will do the simpler single file import.