Update annotations import
Switched from read_csv to read.csv because read.csv throws a warning but still executes if a column is missing, whereas read_csv fails to import the file. The Common/Count column is only present in some of the ASPIRE expedition annotations on SeaTube. Additionally, the Common/Count column was sometimes importing as a logical, so I set the data type to character. My preference would be to set it to numeric, but I found a few instances of non-numeric text in this column across the ASPIRE expeditions.